+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 ## Unreleased
 
+### Added
+* Added User Agent Integration Headers
+
 ## 2.0.1
 ### Fixed
 * Fixed issue when generating file on entries with a field called `Content Type`